requirements.txt
boto3==1.28.52
botocore==1.31.52
certifi==2022.6.15
click==8.1.3
grpcio==1.47.0
grpcio-tools==1.47.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.2.0
protobuf==3.20.1
python-dateutil==2.8.2
qrcode==7.3.1
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
